refactor(web): move all GraphQL fragments to lib/fragments/
Centralize all 10 section fragment definitions in lib/fragments/ with
a barrel index.ts export. Components re-export their fragment for
backward compatibility. Removes fragment definitions from component
files and deletes the old standalone fragment files from
components/sections/.

Made-with: Cursor
