+++ apps/web/src/lib/fragments/bible-quotes-carousel.ts
@@ -1,0 +1,9 @@
+import { graphql } from "@forge/graphql"
+
+export const bibleQuotesCarouselFragment = graphql(`
+  fragment BibleQuotesCarousel on ComponentSectionsBibleQuotesCarousel
+  @_unmask {
+    id
+    heading
+  }
+`)
