+++ apps/web/src/lib/fragments/cta-section.ts
@@ -1,0 +1,11 @@
+import { graphql } from "@forge/graphql"
+
+export const ctaSectionFragment = graphql(`
+  fragment CTASection on ComponentSectionsCta @_unmask {
+    id
+    ctaHeading: heading
+    body
+    buttonLabel
+    buttonLink
+  }
+`)
