+++ apps/web/src/lib/fragments/promo-banner.ts
@@ -1,0 +1,11 @@
+import { graphql } from "@forge/graphql"
+
+export const promoBannerFragment = graphql(`
+  fragment PromoBanner on ComponentSectionsPromoBanner @_unmask {
+    id
+    promoHeading: heading
+    promoDescription: description
+    intro
+    promoCtaLink: ctaLink
+  }
+`)
