+++ apps/web/src/lib/fragments/section.ts
@@ -1,0 +1,17 @@
+import { graphql } from "@forge/graphql"
+
+export const sectionFragment = graphql(`
+  fragment Section on ComponentSectionsSection @_unmask {
+    id
+    sectionKey
+    backgroundColor
+    backgroundOpacity
+    blurHash
+    sectionContent: content {
+      __typename
+      ... on ComponentSectionsContainer {
+        ...Container
+      }
+    }
+  }
+`)
