+++ apps/web/src/lib/fragments/text-section.ts
@@ -1,0 +1,13 @@
+import { graphql } from "@forge/graphql"
+
+export const textSectionFragment = graphql(`
+  fragment TextSection on ComponentSectionsText @_unmask {
+    id
+    sectionKey
+    heading
+    headingLevel
+    subtitle
+    contentParagraphs
+    textVariant: variant
+  }
+`)
